Safari-609.1.11.1/LayoutTests/fast/text/font-face-api-preinstalled-expected.txt

This test makes sure that preinstalled fonts are never returned from the CSS Font Loading API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.length is 0
PASS result.length is 0
PASS successfullyParsed is true

TEST COMPLETE
